requirements.txt
coverage
Django
djangorestframework
drf-nested-routers
